mingw: Treat MAX_PATH as 248, not 260
Infamous limitation of Windows is the 260 max path length that requires
us to expand path to UNC (//?/) in order to cooperate with WinAPI.
Currently, we decide NOT to expand the path if the length of whole path
(base path + relative directory + filename) would not exceed 260.
Unfortunatelly, during creation of directory (base path + relative dir)
if we go over 248 we will trigger another limitation of Windows and
cause _wmkdir to fail.

In order not to complicate logic in handle_long_path (to detect whether
we are a directory and thus we need to use a different constant) we
limit ourselves to 248, always. It should have minimal impact on
performance as usually paths are either short, or pretty long and
expansion would happen anyway.

Signed-off by: Bartosz Bielecki <bielecki.b@gmail.com>